+++ python/tests/conftest.py
@@ -40,6 +40,9 @@ def pytest_configure(config):
     logging.getLogger("semantic_kernel").setLevel(logging.INFO)
 
 
+# region: Kernel fixtures
+
+
 @fixture(scope="function")
 def kernel() -> "Kernel":
     from semantic_kernel import Kernel
@@ -192,6 +195,7 @@ def prompt() -> str:
     return "test prompt"
 
 
+# region: Connector Settings fixtures
 @fixture
 def exclude_list(request):
     """Fixture that returns a list of environment variables to exclude."""
@@ -204,6 +208,7 @@ def override_env_param_dict(request):
     return request.param if hasattr(request, "param") else {}
 
 
+# These two fixtures are used for multiple things, also non-connector tests
 @fixture()
 def azure_openai_unit_test_env(monkeypatch, exclude_list, override_env_param_dict):
     """Fixture to set environment variables for AzureOpenAISettings."""
@@ -271,264 +276,8 @@ def openai_unit_test_env(monkeypatch, exclude_list, override_env_param_dict):
     return env_vars
 
 
-@fixture()
-def mistralai_unit_test_env(monkeypatch, exclude_list, override_env_param_dict):
-    """Fixture to set environment variables for MistralAISettings."""
-    if exclude_list is None:
-        exclude_list = []
-
-    if override_env_param_dict is None:
-        override_env_param_dict = {}
-
-    env_vars = {
-        "MISTRALAI_CHAT_MODEL_ID": "test_chat_model_id",
-        "MISTRALAI_API_KEY": "test_api_key",
-        "MISTRALAI_EMBEDDING_MODEL_ID": "test_embedding_model_id",
-    }
-
-    env_vars.update(override_env_param_dict)
-
-    for key, value in env_vars.items():
-        if key not in exclude_list:
-            monkeypatch.setenv(key, value)
-        else:
-            monkeypatch.delenv(key, raising=False)
-
-    return env_vars
-
-
-@fixture()
-def anthropic_unit_test_env(monkeypatch, exclude_list, override_env_param_dict):
-    """Fixture to set environment variables for AnthropicSettings."""
-    if exclude_list is None:
-        exclude_list = []
-
-    if override_env_param_dict is None:
-        override_env_param_dict = {}
-
-    env_vars = {"ANTHROPIC_CHAT_MODEL_ID": "test_chat_model_id", "ANTHROPIC_API_KEY": "test_api_key"}
-
-    env_vars.update(override_env_param_dict)
-
-    for key, value in env_vars.items():
-        if key not in exclude_list:
-            monkeypatch.setenv(key, value)
-        else:
-            monkeypatch.delenv(key, raising=False)
-
-    return env_vars
-
-
-@fixture()
-def aca_python_sessions_unit_test_env(monkeypatch, exclude_list, override_env_param_dict):
-    """Fixture to set environment variables for ACA Python Unit Tests."""
-    if exclude_list is None:
-        exclude_list = []
-
-    if override_env_param_dict is None:
-        override_env_param_dict = {}
-
-    env_vars = {
-        "ACA_POOL_MANAGEMENT_ENDPOINT": "https://test.endpoint/",
-    }
-
-    env_vars.update(override_env_param_dict)
-
-    for key, value in env_vars.items():
-        if key not in exclude_list:
-            monkeypatch.setenv(key, value)
-        else:
-            monkeypatch.delenv(key, raising=False)
-
-    return env_vars
-
-
-@fixture()
-def azure_ai_search_unit_test_env(monkeypatch, exclude_list, override_env_param_dict):
-    """Fixture to set environment variables for ACA Python Unit Tests."""
-    if exclude_list is None:
-        exclude_list = []
-
-    if override_env_param_dict is None:
-        override_env_param_dict = {}
-
-    env_vars = {
-        "AZURE_AI_SEARCH_API_KEY": "test-api-key",
-        "AZURE_AI_SEARCH_ENDPOINT": "https://test-endpoint.com",
-        "AZURE_AI_SEARCH_INDEX_NAME": "test-index-name",
-    }
-
-    env_vars.update(override_env_param_dict)
-
-    for key, value in env_vars.items():
-        if key not in exclude_list:
-            monkeypatch.setenv(key, value)
-        else:
-            monkeypatch.delenv(key, raising=False)
-
-    return env_vars
-
-
-@fixture()
-def mongodb_atlas_unit_test_env(monkeypatch, exclude_list, override_env_param_dict):
-    """Fixture to set environment variables for MongoDB Atlas Unit Tests."""
-    if exclude_list is None:
-        exclude_list = []
-
-    if override_env_param_dict is None:
-        override_env_param_dict = {}
-
-    env_vars = {"MONGODB_ATLAS_CONNECTION_STRING": "mongodb://test", "MONGODB_ATLAS_DATABASE_NAME": "test-database"}
-
-    env_vars.update(override_env_param_dict)
-
-    for key, value in env_vars.items():
-        if key not in exclude_list:
-            monkeypatch.setenv(key, value)
-        else:
-            monkeypatch.delenv(key, raising=False)
-
-    return env_vars
-
-
-@fixture()
-def bing_unit_test_env(monkeypatch, exclude_list, override_env_param_dict):
-    """Fixture to set environment variables for BingConnector."""
-    if exclude_list is None:
-        exclude_list = []
-
-    if override_env_param_dict is None:
-        override_env_param_dict = {}
-
-    env_vars = {
-        "BING_API_KEY": "test_api_key",
-        "BING_CUSTOM_CONFIG": "test_org_id",
-    }
-
-    env_vars.update(override_env_param_dict)
-
-    for key, value in env_vars.items():
-        if key not in exclude_list:
-            monkeypatch.setenv(key, value)
-        else:
-            monkeypatch.delenv(key, raising=False)
-
-    return env_vars
-
-
-@fixture()
-def google_search_unit_test_env(monkeypatch, exclude_list, override_env_param_dict):
-    """Fixture to set environment variables for the Google Search Connector."""
-    if exclude_list is None:
-        exclude_list = []
-
-    if override_env_param_dict is None:
-        override_env_param_dict = {}
-
-    env_vars = {
-        "GOOGLE_SEARCH_API_KEY": "test_api_key",
-        "GOOGLE_SEARCH_ENGINE_ID": "test_id",
-    }
-
-    env_vars.update(override_env_param_dict)
-
-    for key, value in env_vars.items():
-        if key not in exclude_list:
-            monkeypatch.setenv(key, value)
-        else:
-            monkeypatch.delenv(key, raising=False)
-
-    return env_vars
-
-
-@fixture
-def postgres_unit_test_env(monkeypatch, exclude_list, override_env_param_dict):
-    """Fixture to set environment variables for Postgres connector."""
-    if exclude_list is None:
-        exclude_list = []
-
-    if override_env_param_dict is None:
-        override_env_param_dict = {}
-
-    env_vars = {"POSTGRES_CONNECTION_STRING": "host=localhost port=5432 dbname=postgres user=testuser password=example"}
-
-    env_vars.update(override_env_param_dict)
-
-    for key, value in env_vars.items():
-        if key not in exclude_list:
-            monkeypatch.setenv(key, value)
-        else:
-            monkeypatch.delenv(key, raising=False)
-
-    return env_vars
-
-
-@fixture
-def qdrant_unit_test_env(monkeypatch, exclude_list, override_env_param_dict):
-    """Fixture to set environment variables for QdrantConnector."""
-    if exclude_list is None:
-        exclude_list = []
-
-    if override_env_param_dict is None:
-        override_env_param_dict = {}
-
-    env_vars = {"QDRANT_LOCATION": "http://localhost:6333"}
-
-    env_vars.update(override_env_param_dict)
-
-    for key, value in env_vars.items():
-        if key not in exclude_list:
-            monkeypatch.setenv(key, value)
-        else:
-            monkeypatch.delenv(key, raising=False)
-
-    return env_vars
-
-
-@fixture
-def redis_unit_test_env(monkeypatch, exclude_list, override_env_param_dict):
-    """Fixture to set environment variables for Redis."""
-    if exclude_list is None:
-        exclude_list = []
-
-    if override_env_param_dict is None:
-        override_env_param_dict = {}
-
-    env_vars = {"REDIS_CONNECTION_STRING": "redis://localhost:6379"}
-
-    env_vars.update(override_env_param_dict)
-
-    for key, value in env_vars.items():
-        if key not in exclude_list:
-            monkeypatch.setenv(key, value)
-        else:
-            monkeypatch.delenv(key, raising=False)
-
-    return env_vars
-
-
-@fixture
-def pinecone_unit_test_env(monkeypatch, exclude_list, override_env_param_dict):
-    """Fixture to set environment variables for Pinecone."""
-    if exclude_list is None:
-        exclude_list = []
-
-    if override_env_param_dict is None:
-        override_env_param_dict = {}
-
-    env_vars = {"PINECONE_API_KEY": "test_key"}
-
-    env_vars.update(override_env_param_dict)
-
-    for key, value in env_vars.items():
-        if key not in exclude_list:
-            monkeypatch.setenv(key, value)
-        else:
-            monkeypatch.delenv(key, raising=False)
-
-    return env_vars
-
-
+# region: Data Model Fixtures
+# some of these fixtures are used in both unit and integration tests
 @fixture
 def index_kind(request) -> str:
     if hasattr(request, "param"):
